move ModifierId to scorex-util; add scorex-util dependency;
