+++ build.sbt
@@ -86,7 +86,8 @@ val testingDependencies = Seq(
 
 libraryDependencies ++= Seq(
   "com.iheart" %% "ficus" % "1.4.2",
-  "org.scorexfoundation" %% "scrypto" % "2.+"
+  "org.scorexfoundation" %% "scrypto" % "2.+",
+  "org.scorexfoundation" %% "scorex-util" % "0.1.1-SNAPSHOT"
 ) ++ networkDependencies ++ apiDependencies ++ loggingDependencies ++ testingDependencies
