+++ examples/src/main/scala/examples/hybrid/history/HistoryStorage.scala
@@ -5,7 +5,7 @@ import examples.commons.idToBAW
 import examples.hybrid.blocks._
 import examples.hybrid.mining.{HybridMiningSettings, PosForger}
 import io.iohk.iodb.{ByteArrayWrapper, LSMStore}
-import scorex.core.{ModifierId, bytesToId, idToBytes}
+import scorex.util._
 import scorex.core.consensus.ModifierSemanticValidity
 import scorex.core.consensus.ModifierSemanticValidity.{Absent, Unknown}
 import scorex.core.utils.ScorexLogging
