+++ examples/src/main/scala/examples/hybrid/history/HybridHistory.scala
@@ -16,8 +16,9 @@ import scorex.core.settings.ScorexSettings
 import scorex.core.transaction.box.proposition.PublicKey25519Proposition
 import scorex.core.utils.{NetworkTimeProvider, ScorexEncoding, ScorexLogging}
 import scorex.core.validation.RecoverableModifierError
-import scorex.core.{ModifierId, ModifierTypeId, NodeViewModifier}
+import scorex.core.{ModifierTypeId, NodeViewModifier}
 import scorex.crypto.hash.Blake2b256
+import scorex.util.ModifierId
 
 import scala.annotation.tailrec
 import scala.util.{Failure, Success, Try}
