+++ src/main/scala/scorex/ObjectGenerators.scala
@@ -9,8 +9,9 @@ import scorex.core.network.message.BasicMsgDataTypes._
 import scorex.core.serialization.Serializer
 import scorex.core.transaction.box.proposition.PublicKey25519Proposition
 import scorex.core.transaction.state.{PrivateKey25519, PrivateKey25519Companion}
-import scorex.core.{ModifierId, ModifierTypeId, NodeViewModifier, bytesToId}
+import scorex.core.{ModifierTypeId, NodeViewModifier}
 import scorex.crypto.signatures.Curve25519
+import scorex.util._
 
 import scala.util.Try
