+++ src/main/scala/scorex/core/ModifiersCache.scala
@@ -3,6 +3,7 @@ package scorex.core
 import scorex.core.consensus.HistoryReader
 import scorex.core.utils.ScorexLogging
 import scorex.core.validation.RecoverableModifierError
+import scorex.util.ModifierId
 
 import scala.annotation.tailrec
 import scala.collection.mutable
