+++ src/main/scala/scorex/core/consensus/History.scala
@@ -3,6 +3,7 @@ package scorex.core.consensus
 import scorex.core._
 import scorex.core.consensus.History.ProgressInfo
 import scorex.core.utils.ScorexEncoder
+import scorex.util.ModifierId
 
 import scala.util.Try
