+++ src/main/scala/scorex/core/consensus/HistoryReader.scala
@@ -1,6 +1,7 @@
 package scorex.core.consensus
 
 import scorex.core._
+import scorex.util.ModifierId
 
 import scala.util.Try
