+++ src/main/scala/scorex/core/core.scala
@@ -3,6 +3,7 @@ package scorex
 import scorex.core.network.message.BasicMsgDataTypes.InvData
 import scorex.core.utils.ScorexEncoder
 import scorex.crypto.encode.Base16
+import scorex.util._
 import supertagged.TaggedType
 
 package object core {
@@ -10,13 +11,9 @@ package object core {
   //TODO implement ModifierTypeId as a trait
   object ModifierTypeId extends TaggedType[Byte]
 
-  object ModifierId extends TaggedType[String]
-
   object VersionTag extends TaggedType[String]
 
   type ModifierTypeId = ModifierTypeId.Type
-
-  type ModifierId = ModifierId.Type
 
   type VersionTag = VersionTag.Type
 
@@ -32,11 +29,6 @@ package object core {
   }
 
   def idsToString(invData: InvData)(implicit encoder: ScorexEncoder): String = idsToString(invData._1, invData._2)
-
-
-  def bytesToId(bytes: Array[Byte]): ModifierId = ModifierId @@ Base16.encode(bytes)
-
-  def idToBytes(id: ModifierId): Array[Byte] = Base16.decode(id).get
 
   def bytesToVersion(bytes: Array[Byte]): VersionTag = VersionTag @@ Base16.encode(bytes)
